requirements.txt
asgiref
certifi
cffi
charset-normalizer
crispy-bootstrap5
cryptography
defusedxml
Django
django-allauth
django-crispy-forms
django_debug_toolbar
gunicorn
idna
oauthlib
psycopg2-binary
pycparser
PyJWT
python3-openid
requests
requests-oauthlib
typing_extensions
urllib3
whitenoise
coderedcms
environs
dj-database-url
wagtail-2fa